requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
watchdog>=3.0.0
pyjwt>=2.8.0
requests>=2.31.0
pydantic>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
cryptography>=3.4.0
pyyaml>=6.0.0
python-dotenv>=1.0.0